Updates the most recent catch time series in the SS3 data file

Integrate the latest catch information from the U.S. and Canadian Pacific Hake fisheries into the Stock Synthesis data file. Information on catches is found in .csv files saved in the hake-assessment repository.

commit_catch(
  dir_data = hakedata_wd(),
  digits = 5,
  lastassessmentvals = c(NA, NA, NA)
)

Arguments

dir_data

A directory that houses the .csv data files for the stock assessment. These are typically found within the data directory of the hake-assessment repository.

digits

The number of digits you want the results to saved to. [round](..., digits = digits) will be used.

lastassessmentvals

A vector of three values,

estimate of spawning stock biomass for terminal year in thousands of metric tons,

estimate of depletion in percent with one decimal place,

estimate of TAC for the first year of the forecast period in metric tons rounded to the nearest whole number.

Value

The following two files are updated in the pacific-hake/hake-assessment repository:

landings-tac-history.csv

catch-targets-biomass.csv

Author

Kelli F. Johnson