Plot a box plot with user-defined probabilities

Plot a box plot with user-defined probabilities using ggplot2::geom_boxplot and ggplot2::facet_wrap if more than one xvar is supplied.

plot_boxplot(
  data,
  xvar,
  yvar,
  file,
  probabilities = c(0.025, 0.25, 0.5, 0.75, 0.975),
  showmedian = FALSE,
  mlab = "Box plot with 95 percent quantiles",
  xlab = tools::toTitleCase(xvar[1]),
  ylab = "",
  yscale = "identity",
  incolor,
  scales = "fixed",
  nrow = NULL,
  ncol = NULL,
  legend.position,
  legend.direction = "vertical",
  ...
)

Arguments

data

A data frame.

xvar

A vector of column names available in data.

yvar

A character value representing the column name in data that is the dependent variable. There can only be one entry in this argument, not a vector, and it must match an existing column exactly.

file

A file path for either an existing file or a file that will be written to the disk by the given function. Paths can be either full or relative.

probabilities

A vector of five probabilities, starting with the lower limit of the whisker, then the lower edge of the box, the middle of the box, the upper edge of the box, and the upper limit of the whisker.

showmedian

A logical, specifying if a line should be plotted for the third entry in probabilities, which will be the middle of the box. The default, FALSE, leads to no median/middle line.

mlab

The main title, located above the figure, which will be centered. Please use sentence case without a full stop. To create a new line use \n in the text.

xlab

A label for the x axis. The default is to use the x-axis variable in sentence case without a full stop. Remember to give units if applicable.

ylab

A label for the y axis. The default label is typically empty so users are reminded to explain their y-axis label rather than just using a column name. Please use sentence case without a full stop and units should be given if applicable.

yscale

The desired scale for the y axis. Allowable values are all of those accepted in the trans argument of ggplot2::scale_y_continuous. The default, identity, leads to nothing being changed. Users will typically want to change the value to log, which doesn't transform the data just the distance between the labels on the y axis.

incolor

If missing, no colors will be used for aes(fill = ) in the boxplots. Otherwise, a vector of column names can be used, which will lead to interaction being called and the resulting factor will determine the color categories. Colors will be determined using plot_colour; but, note that this internal function can only handle a limited number of categories.

scales

A character value supplied to ggplot2::facet_wrap that defines how the scales of the axes are determined across facets. The default of fixed causes all axes to be the same across panels. The most-likely value besides the default will be free allowing axes to be specific to each panel.

nrow, ncol

Number of rows and columns passed to ggplot2::facet_wrap. The default values of NULL allow the function to determine the appropriate number of rows and columns. If changed from the default, use integer values to predetermine the number of rows or columns.

legend.position

Coordinates for where you would like the legend to be placed.

legend.direction

"vertical" or `"horizontal", where the default is the former.

...

Arguments passed to ggplot2::ggsave other than filename which is passed using file.

Value

A ggplot2 object is invisibly returned. If file is supplied, then the figure is also saved based on the entry (i.e., either relative to your current working directory or based on the full path if specified). The returned object contains a data frame of the data used in the figure and not the data frame supplied to the figure. This object returned[["data"]]

is helpful when trying to determine quantiles for text in a document.

Examples

testdata <- data.frame(
  yvar = rnorm(100),
  xvar1 = rep(c("Good", "Bad", "Bad", "Good"), length.out = 100),
  xvar2 = rep(c("Apples", "Bananas", "Oranges", "Carrots"), length.out = 100)
)
gg <- plot_boxplot(
  data = testdata, xvar = c("xvar1"), yvar = "yvar",
  showmedian = TRUE, ylab = "Best fruit", incolor = "xvar2",
  legend.position = "right"
)
#> Error in plot_boxplot(data = testdata, xvar = c("xvar1"), yvar = "yvar",     showmedian = TRUE, ylab = "Best fruit", incolor = "xvar2",     legend.position = "right"): could not find function "plot_boxplot"
if (FALSE) {
print(gg)
grDevices::dev.off()
rm(testdata, gg)
}