Process NORPAC catches

Summarize and plot NORPAC catches for the at-sea hake fishery. Fleet types are assigned to create summaries by fleet and year.

process_catch_norpac(
  ncatch = get_local(file = "norpac_catch.Rdat"),
  nyears = 5,
  savedir = hakedata_wd()
)

Arguments

ncatch

A data frame with catch information extracted from the NORPAC database. The default is to read the .Rdat file from the disk using get_local(). Else, users can just pass the R object itself, which would typically be hakedata[["ncatch"]].

nyears

The number of years for plotting.

Value

Figures and csv files are saved to the disk regarding catch rates and depth (fishing and bottom depths). In the Figures directory, the following png files are saved:

fishDepthsUS.png

fishDepthsUSallyears.png

fish_FISHING_DEPTH_FATHOMS_US.png

fish_BOTTOM_DEPTH_FATHOMS_US.png

fishBottomDepthByVesselUS.png

fishCatchRatesUSByYear.png

fishDepthByYearUS.png

../doc/main-figures/fishCatchRatesUS.png

fishCatchRatesUSnolog.png

In the Catches directory, the following csv files are saved:

depth-us-atsea-bottom.csv

depth-us-atsea-fishing.csv

Tribal catches

The column CDQ_CODE contains strings starting with an upper-case followed by and the last two digits of the the CDQ code of the vessel. The code was created in 1995 and is used to track tribal catches in the hake fishery. The following codes exist:

C51 -- Aleutian Pribilof Island Community Development Association

C52 -- Bristol Bay Economic Development Corporation

C53 -- Central Bering Sea Fishermen's Association

C54 -- Coastal Villages Fishing Cooperative

C55 -- Norton Sound Economic Development Corporation

C56 -- Yukon Delta Fisheries Development Association

R10 -- Research Permit - Alaska Fisheries Science Center Catch Estimation

M01 -- Makah Tribe Whiting Association

R11 -- Research Permit - WA Sea Grant bird/longline interaction

R12 -- Research Permit - WA Sea Grant bird/longline interaction

Raw data

Raw data consists of sampled and unsampled hauls with the total weight including bycatch weight. For sampled hauls, EXTRAPOLATED_WEIGHT is the weight of the catch only. For unsampled hauls that have no knowledge of the amount of bycatch included in the haul, OFFICIAL_TOTAL_CATCH must be used along with a estimated bycatch rate to calculate EXTRAPOLATED_WEIGHT. Monthly fleet-specific bycatch rates are estimated from the sampled hauls and multiplied by the OFFICIAL_TOTAL_CATCH to get an estimate of the amount of bycatch that should be subtracted from OFFICIAL_TOTAL_CATCH to get just the weight of Pacific Hake.

Definitions

OFFICIAL_TOTAL_CATCH is measured in metric tons (mt) and includes both retained and discarded species. Thus, this measurement sums the targeted species, prohibited species, and non-allocated species.

EXTRAPOLATED_WEIGHT is measured in kilograms (kg) and includes just the weight of the relevant SPECIES included in that row of data, e.g., if SPECIES == 206, then EXTRAPOLATED_WEIGHT is the kg of Pacific Hake in that haul.

Assumptions

sampled hauls have an entry in the SPECIES column, i.e., hake are denoted with 206

an unsampled haul is represented by one unique record because there no additional lines of catch for the bycatch species given it was not sampled

Author

Kelli F. Johnson