Write a set of Stock Synthesis input files to a directory for bridging

Write a set of Stock Synthesis input files to a directory after modifying them with new data for the year. Children functions manipulates specific parts of the data or control file and the wrapper function ensures that they are written in the correct order and saved using a standard naming protocol.

write_bridging(dir_input, dir_output)

Arguments

dir_input

A string that specifies the path to a directory that stores the input files for a model that you want to use as the basis of the bridging. It will be passed to r4ss::SS_read(). The path will typically be to the previous year's base model, which represents the first model in a bridging set.

dir_output

A string that specifies the path to a directory that may or may not already exist that will be used to save the output.

Value

An invisible list of modified Stock Synthesis input files. A directory is also created in dir_output with multiple directories, each storing a bridging step.

Details

Catch: The catches are taken from data/landings-tac-history.csv, which is updated in process_catch().

Weight-At-Age: The weight-at-age information is stored in data/wtatage.ss, which is updated in process_weight_at_age().

CPUE: Catch-per-unit-effort data for the age-2-plus and age-1 surveys, with the following columns:

year

seas

index

obs

se_log. TODO: needs to be saved somewhere in the repository, this could be in survey-history.csv but the decimal places are not enough right now and the age-1 index is not saved. Would also need to add 0's for the years without a survey.

agecomp_survey Age-composition data for the age-2-plus survey. TODO: save this information somewhere

Author

Kelli F. Johnson